cree las clases de conexionConTablaPagos y ConexionConRecibos para separar todo muy bien, y elimine los metodos existentes en conexionSQL para q estuvieran donde debian estar
