+++ CemprodecaDataManager/src/conexionesBD/conexionSQL.java
@@ -28,21 +28,6 @@ public class conexionSQL {
 	      }
 	}
 	
-	public ArrayList<TablaDePagos> DescargaTablaDePagos(){
-	System.out.println("NO ESTA DEFINIDO EL CONTENIDO DEL METODO");
-	return null;/*retorna un arraylist sacado de la db con las filas de la tabla de pagos*/
-	}/*fin de TablaDePagos entrante*/
+	
 
-	public ArrayList<Recibos> DescargaRecibos(){
-		System.out.println("NO ESTA DEFINIDO EL CONTENIDO DEL METODO");
-	return null;/*retorna los recibos viejos de la db*/
-	}/*fin de Recibos entrantes*/
-	
-	public void CargarTablaDePagosAlaDB(){
-		
-	}/*fin de carga de tabla de pagos*/
-	
-	public void CargarRecibosAlaDB(){
-		
-	}/*fin de carga de recibos*/
 }
